# Conflicts:
#	src/javascript/src/dom_binding.cpp
#	src/media/include/neko/media/media_source.h
#	src/pdf/include/neko/pdf/pdf.h
#	src/pdf/src/parser.cpp
#	src/ui/src/web_view.cpp
#	tests/unit/media/media_test.cpp
#	tests/unit/pdf/pdf_test.cpp

+++ src/renderer/include/neko/renderer/page.h
@@ -29,14 +29,14 @@ namespace neko::renderer {
 // |border_top|/|border_left| the top/left border widths.
 struct ElementGeometry
 {
-  double x = 0;
-  double y = 0;
-  double width = 0;
-  double height = 0;
-  double client_width = 0;
-  double client_height = 0;
-  double border_top = 0;
-  double border_left = 0;
+  float x = 0;
+  float y = 0;
+  float width = 0;
+  float height = 0;
+  float client_width = 0;
+  float client_height = 0;
+  float border_top = 0;
+  float border_left = 0;
 };
 
 // The minimal page pipeline: HTML -> DOM -> style -> layout -> paint.
@@ -236,10 +236,10 @@ private:
   struct ImageAnimationState
   {
     std::shared_ptr<image::GifAnimation> animation;
-    double start_ms = 0;    // steady clock of the first display
-    std::size_t frame = 0;  // currently displayed frame index
-    std::size_t loops = 0;  // completed full passes
-    bool finished = false;  // loop count reached; stays on the last frame
+    double start_ms = 0;   // steady clock of the first display
+    std::size_t frame = 0; // currently displayed frame index
+    std::size_t loops = 0; // completed full passes
+    bool finished = false; // loop count reached; stays on the last frame
   };
   std::unordered_map<const dom::Element*, ImageAnimationState> animation_states_;
 
@@ -251,9 +251,9 @@ private:
     std::shared_ptr<std::vector<image::Image>> frames;
     double frame_rate = 0;
     bool loop = false;
-    bool playing = false;  // autoplay starts it on the first advance
+    bool playing = false; // autoplay starts it on the first advance
     bool autoplay = false;
-    double start_ms = 0;   // steady clock when playback (re)started
+    double start_ms = 0;    // steady clock when playback (re)started
     double paused_time = 0; // playback position when paused
     std::size_t frame = 0;  // currently displayed frame index
   };
